pyproject.toml
[tool.poetry]
name = "passport_detect"
version = "0.1.0"
description = "The passport detection"
authors = ["Lesnoy Elf"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = "^1.26.0"
pandas = "^2.1.1"
requests = "^2.31.0"
scikit-learn = "^1.3.1"
catboost = "^1.2.2"
imblearn = "^0.0"
pre-commit = "^3.4.0"
mlflow = "^2.8.1"
onnxruntime = "^1.16.3"
torch = {version = "^2.1.1+cpu", source = "pytorch"}
torchvision = {version = "^0.16.1+cpu", source = "pytorch"}
tf2onnx = "^1.15.1"
skl2onnx = "^1.15.0"
onnx = "^1.15.0"
dvc = "^3.30.1"
dvc-gdrive = "^2.20.0"
torchmetrics = "^1.2.0"
hydra-core = "^1.3.2"
lightning = "^2.1.2"
pyfunc = "^0.1"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
